Try to extract encoder into `Encoder\*` classes.
